Streamline dots and their positions and more.
Previously, I had a different square size for triple dots. It is
now replaced with the standard 64x64 square everywhere there is a
single, double or triple dot. No exceptions for base glyphs atm.

Moreover, non-essential references are removed to reduce the number of
glyphs.

Some other minor changes are made, for examplt adding a squre to the
bottom-left of Arabic letter HEH final form. Since the font is still
in beta, I'm not going to add them to FONTLOG.

Some one-off mistakes where fixed.