feat: rescale to 100ptx100pt base dot-glyph.
In order to adjust the font size and occuply more of the base
square and the make the font look larger (it was very small),
I had to resize all the glyphs. The base square is now 100x100
points. All other glyphs are adjusted accordingly.

The descent was set to 400 and ascent to 624 making the Em Size
1024 points.